+++ lib/weird_phlex.rb
@@ -2,4 +2,6 @@
 
 require 'weird_phlex/version'
 require 'weird_phlex/railtie' if defined?(Rails::Railtie)
-require 'weird_phlex/core'
+
+require 'weird_phlex/indexer'
+require 'weird_phlex/planner'
