--- lib/weird_phlex/core.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require 'weird_phlex/core/indexer'
-require 'weird_phlex/core/planner'
