--- lib/weird_phlex/core/component_pack/component.rb
@@ -1,20 +1,0 @@
-# frozen_string_literal: true
-
-module WeirdPhlex
-  module Core
-    module ComponentPack
-      class Component
-        attr_reader :files
-
-        def initialize(name, files: [])
-          @name = name
-          @files = files
-        end
-
-        def to_s
-          "#{@name} - #{@files.count} file#{'s' if @files.count != 1}"
-        end
-      end
-    end
-  end
-end
